Refactor: Relocate UI components to `library` package
This commit moves UI-related files from the `com.mshdabiola.ui` package to `com.mshdabiola.library` within the `library` module.

- Moved `NoteCard.kt`, `Note.kt`, and `NoteCardTestTags.kt` from `library/src/commonMain/kotlin/com/mshdabiola/ui` to `library/src/commonMain/kotlin/com/mshdabiola/library`.
- Moved `NoteCardTest.kt` from `library/src/jvmTest/kotlin/com/mshdabiola/ui` to `library/src/jvmTest/kotlin/com/mshdabiola/library`.
- Updated package declarations in the moved files.
- Removed an unused import `org.jetbrains.compose.ExperimentalComposeLibrary` and a comment from `library/build.gradle.kts`.
